backend/requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI Models
torch>=2.0.0
torchaudio>=2.0.0
# torchcodec>=0.8.0  # Disabilitato: richiede FFmpeg e può causare errori
openai-whisper>=20231117
faster-whisper>=1.0.0  # 4-10x piu' veloce, stessa qualita'
# demucs>=4.0.0  # Disabilitato: richiede torchcodec/FFmpeg
openai>=1.3.0
transformers>=4.30.0  # Per modelli locali GPT-2
# spleeter==2.1.0  # Non compatibile con Python 3.14 - usa fallback manuale

# Audio processing
numpy>=1.24.0
pydub>=0.25.1
soundfile>=0.12.0
librosa>=0.11.0
crepe>=0.0.14  # Per pitch detection avanzata
scipy>=1.10.0  # Per filtri frequenza (separazione vocale)
noisereduce>=2.0.0  # Per denoise vocale professionale
# essentia>=2.1b6  # Opzionale: analisi avanzata (richiede build complesso)

# Metricizzazione e sillabe (stile Beatles)
pronouncing>=0.2.0  # CMUdict per conteggio sillabe e accenti
tqdm>=4.65.0  # Progress bar

# Utilities
python-dotenv==1.0.0
pydantic-settings>=2.0.0  # Per configurazione centralizzata
requests>=2.31.0  # Per Ollama API

# Memory Video (opzionale - per salvare testi generati)
memvid>=0.1.3  # Salva testi in video MP4 con ricerca semantica

